+++ changelog.md
@@ -3,6 +3,7 @@ Changelog
 
 1.2.0-dev - IN PROGRESS
 ---------
++ Add `angle.log`
 + Add `point.pointAtBisector`
 + Add `ray.bezierArc`
 + Add `segment.inverse`