+++ changelog.md
@@ -14,6 +14,7 @@ Changelog
 + Modify `Text.Format` constructor defaults and parameter order
 + Modify `Text.Format` to allow nullable `size` and `font`
 + Fix incorrect values set in `rac.Text.Format.topLeft` and `rac.Text.Format.topRight`
++ Modify `rac.Text` to allow an optional `format` parameter
 + Modify `point.text` to allow an optional `format` parameter
 + Rename `Text.Format.horizontal` to `horizontalAlign`
 + Rename `Text.Format.vertical` to `verticalAlign`