+++ changelog.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.2.1-dev - IN PROGRESS
+-----------------------
++ Update `minami-rac` to `1.4.0`
+
+
 1.2.0 - 2022 September 8th
 --------------------------
 Streamline and document Text and Text.Format